+++ WobbleAnimation.cpp
@@ -14,7 +14,7 @@ m_maxHeight (maxHeight)
 void WobbleAnimation::Tick()
 {
 	auto timeValue = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-	auto sinTime = fabs(sin(timeValue * 0.001));
+	auto sinTime = (float)fabs(sin(timeValue * 0.001));
 	m_shape->SetHeight(sinTime * m_maxHeight);
 	m_shape->SetWidth(sinTime * m_maxWidth);
 }
